+++ src/Components/Trials/Trials.tsx
@@ -14,11 +14,11 @@ import React, { useCallback, useEffect, useState } from 'react'
 import { useDropzone } from 'react-dropzone'
 
 import { getFileExtension } from '../../hooks/useAttachment'
+import { readDocumentAsync } from '../../hooks/useAttachmentDropzone'
 import useIntersectionObserver from '../../hooks/useIntersectionObserver'
 import { Trial } from '../../model/model'
 import { FirebaseService } from '../../services/firebase'
 import { useFirebaseAuthContext } from '../Provider/FirebaseAuthProvider'
-import { readDocumentAsync } from '../Recipe/Create/Attachments/useAttachmentDropzone'
 import NotFound from '../Shared/NotFound'
 import Skeletons from '../Shared/Skeletons'
 import TrialsCard from './TrialsCard'
