+++ src/hooks/useAttachmentDropzone.tsx
@@ -3,9 +3,9 @@ import compressImage from 'browser-image-compression'
 import React, { useCallback, useState } from 'react'
 import { DropzoneState, useDropzone } from 'react-dropzone'
 
-import { AttachmentDoc, DataUrl } from '../../../../model/model'
-import { FirebaseService } from '../../../../services/firebase'
-import { useFirebaseAuthContext } from '../../../Provider/FirebaseAuthProvider'
+import { useFirebaseAuthContext } from '../Components/Provider/FirebaseAuthProvider'
+import { AttachmentDoc, DataUrl } from '../model/model'
+import { FirebaseService } from '../services/firebase'
 
 export const readDocumentAsync = (document: Blob) =>
     new Promise<string>((resolve, reject) => {
